Time: 1420 ms (5.03%) | Memory: 53 MB (45.89%) - LeetSync

+++ 169-majority-element/majority-element.java
@@ -9,7 +9,7 @@ class Solution {
                     count++;
                 }
             }
-            if (count > n / 2) {
+            if (count > (int)(n / 2)) {
                 return nums[i];
             }
         }
